refactor: Bot不再实现具体API,增加频道事件以及Api初步支持

+++ config.go
@@ -20,13 +20,13 @@ var defaultData []byte
 //
 type Config struct {
 	NickName         []string `json:"nick_name" yaml:"nick_name" hjson:"nick_name"`
-	Admin            int      `json:"admin" yaml:"admin" hjson:"admin"`
+	Admin            int64    `json:"admin" yaml:"admin" hjson:"admin"`
 	Host             string   `json:"host" yaml:"host" hjson:"host"`
 	Port             int      `json:"port" yaml:"port" hjson:"port"`
 	Token            string   `json:"token"`
 	LogLevel         string   `json:"log_level" yaml:"log_level" hjson:"log_level"`
 	LogTruncate      bool     `json:"log_truncate" yaml:"log_truncate" hjson:"log_truncate"`
-	SuperUser        []int    `json:"super_user" yaml:"super_user" hjson:"super_user"`
+	SuperUser        []int64  `json:"super_user" yaml:"super_user" hjson:"super_user"`
 	CommandStart     []string `json:"command_start" yaml:"command_start" hjson:"command_start"`
 	EnablePlaywright bool     `json:"enable_playwright" yaml:"enable_playwright" hjson:"enable_playwright"`
 	ListenHost       string   `json:"listen_host" yaml:"listen_host"`
